+++ nisystemlink/clients/artifact/models/__init__.py
@@ -1,0 +1,3 @@
+from ._upload_artifact_response import UploadArtifactResponse
+
+# flake8: noqa