+++ nisystemlink/clients/testmonitor/_test_monitor_client.py
@@ -10,6 +10,9 @@ from . import models
 
 
 class TestMonitorClient(BaseClient):
+    # prevent pytest from thinking this is a test class
+    __test__ = False
+
     def __init__(self, configuration: Optional[core.HttpConfiguration]):
         if configuration is None:
             configuration = core.JupyterHttpConfiguration()